IntegrityForge/requirements.txt
PyYAML>=6.0.1
jsonschema>=4.21.1
python-dotenv>=1.0.0